+++ Car_Wash.py
@@ -1,0 +1,7 @@
+class Car_Wash:
+    def __init__(self, car_wash_type):
+        self.car_wash_type = car_wash_type
+
+    def Wash_Car(self):
+        return "Washing car"
+
