+++ Tier_1.py
@@ -1,0 +1,5 @@
+from Car_Wash import Car_Wash
+
+class Tier_1(Car_Wash):
+    def Wash_Car(self):
+        return f"Soap only car wash performed!"
