+++ Tier_2.py
@@ -1,0 +1,5 @@
+from Car_Wash import Car_Wash
+
+class Tier_2(Car_Wash):
+    def Wash_Car(self):
+        return f"Soap and polish car wash performed!"
